apps/amazon-ynab-reconciler/requirements.txt
# Core Dependencies
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=0.21.0
pyotp>=2.8.0
beautifulsoup4>=4.11.0

# Google APIs
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# For testing
pytest>=7.0.0

# Optional: For local Playwright automation (if not using MCP)
# playwright>=1.40.0

# Note: For Lambda deployment, use lambda/requirements.txt which includes:
# - boto3 (for AWS services)
# - playwright (for containerized Lambda)
# - google-auth packages (for email)